sdks/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hotplex"
version = "0.1.0"
description = "Production-ready Python client for HotPlex AI Agent Control Plane"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "HotPlex Team", email = "team@hotplex.dev"}
]
keywords = ["ai", "agent", "claude", "opencode", "control-plane"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "sseclient-py>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/hrygo/hotplex"
Documentation = "https://github.com/hrygo/hotplex#readme"
Repository = "https://github.com/hrygo/hotplex"
Issues = "https://github.com/hrygo/hotplex/issues"

[tool.setuptools.packages.find]
where = ["."]
